검색 내역이 존재하지 않을 경우 화면 구현

+++ AppleAppStore/Model/Search/SearchResult.swift
@@ -51,7 +51,7 @@ struct SearchItemResult: Decodable {
     let artistID: Int
     let artistName: String
     let genres: [String]
-    let price: Int?
+    let price: Double?
     let resultDescription: String
     let userRatingCount: Int
